backend/requirements.txt
Flask==3.1.0
werkzeug==3.1.3
flask-cors==5.0.0
websocket-client==1.2.3